Finshed moving getProjectUrl() to project.js.
