+++ apps/src/code-studio/initApp/project.js
@@ -122,15 +122,43 @@ var projects = module.exports = {
   },
 
   /**
-   * @returns the absolute url to the root of this project without a trailing slash.
-   *     For example: http://studio.code.org/projects/applab/GobB13Dy-g0oK
+   * This method is used so that it can be mocked for unit tests.
    */
-  getProjectUrl() {
-    const match = location.href.match(PROJECT_URL_PATTERN);
+  getUrl() {
+    return location.href;
+  },
+
+  /**
+   * @param [fragment] optional url fragment to append to the end of the project URL.
+   * @returns the absolute url to the root of this project without a trailing slash.
+   *     For example: http://studio.code.org/projects/applab/GobB13Dy-g0oK. Hash strings
+   *     are removed, but query strings are retained. If provided, fragment will be
+   *     added to the end of the URL, before the query string.
+   */
+  getProjectUrl(fragment = '') {
+    const match = this.getUrl().match(PROJECT_URL_PATTERN);
+    let url;
     if (match) {
-      return match[1];
+      url = match[1];
+    } else {
+      url = this.getUrl(); // i give up. Let's try this?
     }
-    return location.href; // i give up. Let's try this?
+    var hashIndex = url.indexOf('#');
+    if (hashIndex !== -1) {
+      url = url.substring(0, hashIndex);
+    }
+    var queryString = '';
+    var queryIndex = url.indexOf('?');
+    if (queryIndex !== -1) {
+      queryString = url.substring(queryIndex);
+      url = url.substring(0, queryIndex);
+    }
+    if (fragment.startsWith('/')) {
+      while (url.endsWith('/')) {
+        url = url.substring(0, url.length - 1);
+      }
+    }
+    return url + fragment + queryString;
   },
 
   getCurrentTimestamp() {
